Move HandleIdentityResult from Repository to Service.
