+++ src/services/Identity/Identity.Domain/Entities/IAppUserRepository.cs
@@ -1,12 +1,13 @@
 using Identity.Domain.Entities;
+using Microsoft.AspNetCore.Identity;
 
 namespace Namespace;
 
 public interface IAppUserRepository
 {
-    Task CreateUserAsync(AppUser appUser, string password);
+    Task<IdentityResult> CreateUserAsync(AppUser appUser, string password);
     Task<IEnumerable<AppUser>> GetAllUsersAsync();
     Task<AppUser?> GetUserByIdAsync(string id);
-    Task DeleteUserAsync(AppUser appUser);
+    Task<IdentityResult> DeleteUserAsync(AppUser appUser);
     Task<bool> IsUserInRoleAsync(AppUser appUser, string role);
 }
